Update es.json
Hello,

I have translated the file into Spanish trying to make it as accurate as possible. This is my first time trying to help, so sorry if I'm not doing it the right way.

I have noted that the // machine names part seems to not work, since I have translated it and the changes do not appear in the game. I have tried to make it work, but I don't have the knowledge to do it.

Thank you very much for reading me and I hope my contribution is useful.

Regards.

JJ_Grud